+++ packages/core/src/components/DropdownNew/components/DropdownWrapperUI.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { DropdownContext } from "../context/DropdownContext";
 import { DropdownContextProps } from "../context/DropdownContext.types";
-import DropdownStructure from "./DropdownStructure/DropdownStructure";
 import DropdownPopup from "./DropdownPopup/DropdownPopup";
 import { BaseListItemData } from "../../BaseListItem";
+import DropdownBase from "./DropdownBase/DropdownBase";
 
 interface DropdownWrapperUIProps<Item extends BaseListItemData<Record<string, unknown>>> {
   contextValue: DropdownContextProps<Item>;
@@ -17,9 +17,9 @@ const DropdownWrapperUI = <Item extends BaseListItemData<Record<string, unknown>
 
   return (
     <DropdownContext.Provider value={contextValue}>
-      <DropdownStructure dropdownRef={dropdownRef}>
+      <DropdownBase dropdownRef={dropdownRef}>
         <DropdownPopup />
-      </DropdownStructure>
+      </DropdownBase>
     </DropdownContext.Provider>
   );
 };
